renamed new data assimilation module SR_DA -> data_assimilation_SR
