+++ igm/processes/iceflow/unified/mappings/interfaces/data_assimilation.py
@@ -30,7 +30,7 @@ class InterfaceDataAssimilation(InterfaceMapping):
     @staticmethod
     def _parse_specs(cfg: DictConfig) -> List[VariableSpec]:
         specs = []
-        for item in cfg.processes.SR_DA.variables:
+        for item in cfg.processes.data_assimilation_SR.variables:
             name = str(item["name"])
             transform = str(item.get("transform", "identity")).lower()
             if transform not in TRANSFORMS.keys():
